API

SpatialHashTables.AbstractSpatialHashTable

SpatialHashTables.BoundedHashTable

SpatialHashTables.BoundedHashTable

SpatialHashTables.BoundedHashTable

SpatialHashTables.BoundedHashTable

SpatialHashTables.BoundedHashTable

SpatialHashTables.BoundedHashTable

SpatialHashTables.BoundedHashTable

SpatialHashTables.SpatialHashTable

SpatialHashTables.SpatialHashTable

SpatialHashTables.SpatialHashTable

SpatialHashTables.SpatialHashTable

SpatialHashTables.SpatialHashTable

SpatialHashTables.SpatialHashTable

Adapt.adapt_structure

Base.resize!

SpatialHashTables.dimension

SpatialHashTables.hashposition

SpatialHashTables.insidegrid

SpatialHashTables.inttype

SpatialHashTables.iterate_box

SpatialHashTables.neighbouring_boxes

SpatialHashTables.neighbours

SpatialHashTables.updatetable!

SpatialHashTables.AbstractSpatialHashTable — Type

AbstractSpatialHashTable

Abstract type for spatial hash tables.

Expected to implement the following functions: dimension(::AbstractSpacialHashTable) inttype(ht::AbstractSpatialHashTable) insidegrid(ht::BoundedHashTable, gridpos) gridindices(ht::BoundedHashTable, pos) hashindex(ht::BoundedHashTable, gridindices)

Moreover, we assume that the hash table has the following fields: ht.cellcount ht.particlemap

SpatialHashTables.BoundedHashTable — Type

BoundedHashTable

A spatial hash table with a fixed domain.

SpatialHashTables.BoundedHashTable — Method

BoundedHashTable(X::AbstractVector, args...)

Constructs a BoundedHashTable with length(X) particles and calls updatetable!(ht, X) to update the hash table.

SpatialHashTables.BoundedHashTable — Method

BoundedHashTable(N::Integer, grid::Tuple, range::AbstractVector)

Constructs a BoundedHashTable with N particles, a grid size of grid and a domain from zero(range) to range. See details of the main constructor for more information.

SpatialHashTables.BoundedHashTable — Method

BoundedHashTable(N::Int64, cellsize::Number, domainstart, domainend)

Same as previous constructor, but allowing the domain to be specified by Vector{Float64} or NTuple{Dim,Float64} instead of StaticVectors. cellsize determines the cell size in each dimension.

SpatialHashTables.BoundedHashTable — Method

BoundedHashTable(N::Int64, cutoff::Number, domainstart::SVector, domainend::SVector)

Constructs a BoundedHashTable with N particles. The cell size is close to cutoff (or slightly larger) in each dimension. The domain ranges from domainstart to domainend.

SpatialHashTables.BoundedHashTable — Method

BoundedHashTable(N::Integer, grid, domainstart, domainend)

Same as the main constructor, but allowing the domain to be specified by Vector{Float64} or NTuple{Dim,Float64} instead of StaticVectors

SpatialHashTables.BoundedHashTable — Method

BoundedHashTable(N::Integer, grid::Tuple, domainstart::SVector, domainend::SVector)

Primary constructor for a BoundedHashTable with N particles, a grid size of grid and a domain from domainstart to domainend. The grid is a tuple of size Dim which determines the number of cells along each dimension.

We refer to the other constructors for more convenient ways to construct a BoundedHashTable.

SpatialHashTables.SpatialHashTable — Type

SpatialHashTable

A spatial hash table for unbounded domains. See Matthias Müller's paper for more information of the method used.

SpatialHashTables.SpatialHashTable — Method

SpatialHashTable(X::AbstractVector, cutoff::Number, tablesize::Int64, args...)

Constructs a SpatialHashTable with length(X) particles and calls updatetable!(ht, X) to update the hash table. The cutoff determines the cell size along each dimension and the Dim is inferred from the first particle in X. (Alternatively one can provide a SVector of size Dim or a NTuple{Dim,Float64}.)

SpatialHashTables.SpatialHashTable — Method

SpatialHashTable(X::AbstractVector, cutoff::Number, tablesize::Int64, args...)

Constructs a SpatialHashTable with length(X) particles and calls updatetable!(ht, X) to update the hash table.

SpatialHashTables.SpatialHashTable — Method

SpatialHashTable(N::Integer, cellsize, tablesize::Integer; args...)

Same as the main constructor, but allowing the cell size to be specified by a tuple of size Dim.

SpatialHashTables.SpatialHashTable — Method

SpatialHashTable(N::Integer, cutoff::Number, Dim::Integer, tablesize::Integer; args...)

Same as the main constructor, but allowing the cell size to be specified by a number cutoff and the spatial dimension Dim.

SpatialHashTables.SpatialHashTable — Method

SpatialHashTable(N::Integer, cellsize::SVector, tablesize::Integer; cachesize=Threads.nthreads(), pseudorandomfactors=default_factors)

Primary constructor for a SpatialHashTable with N particles, a table size of tablesize and a cell size of cellsize. The table size determines the number of cells in the hash table. The cell size is a SVector of size Dim which determines the size of each cell along each dimension.

Additionally, the cachesize determines the number of thread-local caches used for the hash table. Once can set cachesize=1 for single threaded applications.

The pseudorandomfactors are used to generate a hash index for each particle. The default values are taken from Matthias Müller's paper.

Adapt.adapt_structure — Method

Adapt.adapt_structure(to, ht::AbstractSpatialHashTable)

Adapts the hash table ht to the type to. This is useful for GPU support.

It automatically converts types to Float64 and Int32!

Base.resize! — Method

resize!(ht::AbstractSpatialHashTable, n_positions)

Resizes the hash table ht to accomodate n_positions particles. This method is useful if the number of particles changes over time.

This call should usually be followed by a call to updatetable! to update the hash table. Without updatetable! the hash table will be in an inconsistent state.

SpatialHashTables.dimension — Method

dimension(ht::AbstractSpatialHashTable)

Returns the spatial dimension of a hash table.

SpatialHashTables.hashposition — Method

hashposition(ht::AbstractSpatialHashTable, pos)

Returns the hash position of pos in the hash table ht.

SpatialHashTables.insidegrid — Method

inside(ht::AbstractSpatialHashTable, pos)

Returns true if pos is inside the domain of the hash table ht.

SpatialHashTables.inttype — Method

inttype(ht::AbstractSpatialHashTable)

Returns the integer type used for indexing the hash table. (Useful for GPU support.)

SpatialHashTables.iterate_box — Method

iterate_box(ht::AbstractSpatialHashTable, boxhash)

Returns an iterator over the particles in the box with hash boxhash in the hash table ht.

SpatialHashTables.neighbouring_boxes — Method

neighbouring_boxes(ht::AbstractSpatialHashTable, gridpos, r)

Returns an iterator over the neighbouring boxes of the box with grid position gridpos in the hash table ht. r determines the radius of the neighbourhood.

The gridpos is a tuple of size Dim which determines the position of a box in lattice with basis vectors ht.inv_cellsize.

SpatialHashTables.neighbours — Method

neighbours(ht::AbstractSpatialHashTable, pos, r)

Returns an iterator over the particles in the neighbourhood of pos in the hash table ht. r determines the radius of the neighbourhood.

This is the main method of this package and is used to find the neighbours of a particle.

SpatialHashTables.updatetable! — Method

updatetable!(ht::AbstractSpatialHashTable, X)

Updates the hash table ht with the positions X of the particles.

This method is linear in the number of particles and should generally not allocate.

The ideal datatype for X is Vector{SVector{Dim,FT}} where Dim is the spatial dimension and FT is the floating point type.